+++ s4/cli.py
@@ -11,11 +11,31 @@ from inotify_simple import flags
 from s4 import VERSION
 from s4 import sync
 from s4 import utils
+from s4.clients.local import get_local_client
+from s4.clients.s3 import get_s3_client
 from s4.commands.add_command import AddCommand
 from s4.commands.ls_command import LsCommand
 from s4.commands.sync_command import SyncCommand
 from s4.commands.targets_command import TargetsCommand
 from s4.inotify_recursive import INotifyRecursive
+
+
+def get_clients(entry):
+    target_1 = entry['local_folder']
+    target_2 = entry['s3_uri']
+    aws_access_key_id = entry['aws_access_key_id']
+    aws_secret_access_key = entry['aws_secret_access_key']
+    region_name = entry['region_name']
+
+    # append trailing slashes to prevent incorrect prefix matching on s3
+    if not target_1.endswith('/'):
+        target_1 += '/'
+    if not target_2.endswith('/'):
+        target_2 += '/'
+
+    client_1 = get_local_client(target_1)
+    client_2 = get_s3_client(target_2, aws_access_key_id, aws_secret_access_key, region_name)
+    return client_1, client_2
 
 
 def main(arguments):
@@ -133,7 +153,7 @@ def main(arguments):
 
 
 def get_sync_worker(entry):
-    client_1, client_2 = utils.get_clients(entry)
+    client_1, client_2 = get_clients(entry)
     return sync.SyncWorker(client_1, client_2)
